Create experimental file

This function generates replications of mass spectra that are consistent with common experimental data files

create_experimental_file(
  peptides,
  times = c(0.167, 1, 5, 25, 1440),
  charge,
  n_replicates = 3,
  mass_deviations = 50,
  intensity_deviations = NULL,
  file_type = "DynamX"
)

Arguments

peptides

a data frame of sequences (sequence), Protein, and Start, End and parameters except times that can be used for simulating mass spectra. See simulate_theoretical_spectra for more details about the additional parameters.

times

a vector of times at which deuteration levels will be measured (seconds)

charge

vector of charges of the peptide ion. If NULL, one value is sampled from vector 2:6. Default NULL.

n_replicates

number of technical replicates to create

mass_deviations

mass deviation in parts per million. Either a single number (then the error at each time point will be the same) or a vector of the same length as number of unique time points in the experiment. The error will be sampled from normal distribution with standard deviation equal to $$mass_deviations * undeuterated_mass/1e6$$ Default to 50.

intensity_deviations

optional, standard deviations of random noise that will be added to intensities. Either a single number (then the error at each time point will be the same) or a vector of the same length as number of unique time points in the experiment. The error will be sampled from normal distribution with these standard deviations.Default NULL.

file_type

the type of file. Default to "DynamX".

Value

data table. The table of HDX-MS results consistent with `file_type` format.

Examples

peptides <- data.frame(sequence = c("FPTTKTY", "LVRKDLQN"),
                       protection_factor = c(10, 100))
create_experimental_file(peptides, charge = 1:3)
#> [1] "Simulation row: 1"
#> [1] "Simulation row: 2"
#>      Protein Start End Sequence Modification Fragment MaxUptake      MHP  State
#>   1: Protein            FPTTKTY           NA       NA         4 856.4331  PF_10
#>   2: Protein            FPTTKTY           NA       NA         4 856.4331  PF_10
#>   3: Protein            FPTTKTY           NA       NA         4 856.4331  PF_10
#>   4: Protein            FPTTKTY           NA       NA         4 856.4331  PF_10
#>   5: Protein            FPTTKTY           NA       NA         4 856.4331  PF_10
#>  ---                                                                           
#> 104: Protein           LVRKDLQN           NA       NA         6 984.5716 PF_100
#> 105: Protein           LVRKDLQN           NA       NA         6 984.5716 PF_100
#> 106: Protein           LVRKDLQN           NA       NA         6 984.5716 PF_100
#> 107: Protein           LVRKDLQN           NA       NA         6 984.5716 PF_100
#> 108: Protein           LVRKDLQN           NA       NA         6 984.5716 PF_100
#>      Exposure File z RT   Center
#>   1:    0.000    1 1  0 856.4306
#>   2:    0.000    2 1  0 856.4042
#>   3:    0.000    3 1  0 856.4486
#>   4:    0.167    1 1  0 857.2050
#>   5:    0.167    2 1  0 857.1225
#>  ---                            
#> 104:   25.000    2 3  0 985.8800
#> 105:   25.000    3 3  0 986.0011
#> 106: 1440.000    1 3  0 986.0113
#> 107: 1440.000    2 3  0 986.1213
#> 108: 1440.000    3 3  0 986.1768