Get a matrix of simulated exchanged hydrogens for each experiment time point

Calculate matrices of simulated exchange required for obtaining empirical distribution.

get_HD_matrices(
  sequence,
  transition_probs,
  experiment_times,
  times_to_record,
  n_molecules = 100
)

Arguments

sequence
amino acid sequence of a peptide as a character vector

transition_probs
list of probabilities of exchange returned by the get_exchange_probabilities function

experiment_times
numeric vector of times at which exchange will happen

times_to_record
numeric vector of times for which deuteration level measurement should be made

n_molecules
number of peptide molecules

Value

Matrices are stored in a list of matrices (HD_matrices) - each matrix for the respective time point of the measurement times.

Details

At each time point in the time sequence:

change H to D with probability \(P(H -> D)\) in each entry of the matrix from the previous iteration,

change D to H with probability \(P(D -> H)\) in each entry of the matrix from the previous iteration.