Calculate isotopic probabilities (intensity) and mass-to-charge ratio (m/z).

Compute the isotopic probabilities for the deuterated peptide as a convolution of the isotopic distribution for the undeuterated peptide and the observed isotopic distribution after hydrogen-deuterium exchange computed by get_observed_iso_dist.

get_iso_probs_deut(
  HD_matrices,
  maxD,
  maxND,
  isotopic_probs,
  peptide_mass,
  times,
  charge,
  pH
)

Arguments

HD_matrices

list. Simulated matrices for every time point after hydrogen-deuterium-exchange. Calculated via get_HD_matrices or get_HD_matrices_using_markov

maxD

length of the sequence - amount of prolines

maxND

length of the isotopic distribution - 1

isotopic_probs

the isotopic distribution for the undeuterated peptide.

peptide_mass

mass of the peptide + mass of H2O

times

a vector of times at which deuteration levels will be measured (seconds)

charge

vector of charges of the peptide ion. If NULL, one value is sampled from vector 2:6. Default NULL.

pH

pH of the reaction. Default to 7.5.

Value

The output is a data frame with the variables: Exposure (time point of measurement consistent with given HD matrix), Mz - m/z values, Intensity - isotopic probabilities, PH - pH.

Details

The m/z values for the deuterated peptide are calculated using the peptide_mass, charge and constants - deuteron mass (1.00628) and proton mass (1.007276). Starting from the m/z value for the monoisotopic peak, the difference between the mass of deuteron and proton divided by the charge of the peptide ion is added.