Simulate theoretical spectra of a deuterated peptide over time

Simulate theoretical spectra of a deuterated peptide over time

simulate_theoretical_spectra(
  sequence,
  charge = NULL,
  protection_factor = 1,
  times = c(60, 600),
  pH = 7.5,
  temperature = 15,
  n_molecules = 100,
  time_step_const = 1,
  if_corr = FALSE,
  min_probability = 1e-04,
  use_markov = TRUE
)

Arguments

sequence

amino acid sequence of a peptide as a single string

charge

vector of charges of the peptide ion. If NULL, one value is sampled from vector 2:6. Default NULL.

protection_factor

protection factor. If a single number of provided, same protection factor will be assumed for each amide. Default value: 1 (indicates that the exchange rate is equal to the intristic exchange rate)

times

a vector of times at which deuteration levels will be measured (seconds)

pH

pH of the reaction. Default to 7.5.

temperature

temperature of the reaction (Celsius)

n_molecules

number of peptide molecules. Default to 100.

time_step_const

time step constant. Default value: $1$. Value that indicates the length of the time step of the simulation. The bigger the time step, the fewer time points are simulated (the fewer iterations in case of Zhong-Yuan Kan's approach).

if_corr

logical. PH correction indicator. Default value FALSE. The value of pH is equal to pD. If there is correction, the pD = pH + 0.4. (Conelly et al 1993)

min_probability

smallest isotopic probability to consider

use_markov

logical. If TRUE algorithm basing on Markov chain will be used. If FALSE simulation provided by Zhong-Yuan Kan will be executed. Default to TRUE, as it fastens the calculation

Value

a data table of variables:

- Exposure - time point of a measurement,

- Mz - mass-to-charge ratio,

- Intensity - isotopic probabilities larger than

min_probability(the smaller ones are zeroes)

and the variables provided by user

- Sequence,

- PF,

- Charge,

- PH.

Details

To the results calculated by get_iso_probs_deut is added a minimal exchange control - for time point 0 (directly after adding a buffer). The m/z values are obtained as a ratio of the peptide_mass magnified by proton mass and the peptide charge. The distribution of undeuterated peptide is the intensities vector.

See also

Examples

simulate_theoretical_spectra(sequence = "LVRKDLQN",
                             charge = c(3, 5),
                             protection_factor = 100,
                             times = c(0.167, 5),
                             pH = 7.5,
                             temperature = 15,
                             n_molecules = 500,
                             time_step_const = 1,
                             use_markov = TRUE)
#>     Exposure  PH    Intensity       Mz Charge Sequence  PF
#>  1:    0.000 7.5 0.5801405135 329.1978      3 LVRKDLQN 100
#>  2:    0.000 7.5 0.5801405135 197.9216      5 LVRKDLQN 100
#>  3:    0.000 7.5 0.3031692943 329.1978      3 LVRKDLQN 100
#>  4:    0.000 7.5 0.3031692943 197.9216      5 LVRKDLQN 100
#>  5:    0.000 7.5 0.0919265899 329.1978      3 LVRKDLQN 100
#>  6:    0.000 7.5 0.0919265899 197.9216      5 LVRKDLQN 100
#>  7:    0.000 7.5 0.0204531418 329.1978      3 LVRKDLQN 100
#>  8:    0.000 7.5 0.0204531418 197.9216      5 LVRKDLQN 100
#>  9:    0.000 7.5 0.0036689790 329.1978      3 LVRKDLQN 100
#> 10:    0.000 7.5 0.0036689790 197.9216      5 LVRKDLQN 100
#> 11:    0.000 7.5 0.0005576524 329.1978      3 LVRKDLQN 100
#> 12:    0.000 7.5 0.0005576524 197.9216      5 LVRKDLQN 100
#> 13:    0.167 7.5 0.5685853673 329.1978      3 LVRKDLQN 100
#> 14:    0.167 7.5 0.3087345996 329.5332      3 LVRKDLQN 100
#> 15:    0.167 7.5 0.0961595050 329.8687      3 LVRKDLQN 100
#> 16:    0.167 7.5 0.0218844453 330.2041      3 LVRKDLQN 100
#> 17:    0.167 7.5 0.0040049980 330.5395      3 LVRKDLQN 100
#> 18:    0.167 7.5 0.0006199309 330.8749      3 LVRKDLQN 100
#> 19:    0.167 7.5 0.5685853673 197.9216      5 LVRKDLQN 100
#> 20:    0.167 7.5 0.3087345996 198.1229      5 LVRKDLQN 100
#> 21:    0.167 7.5 0.0961595050 198.3241      5 LVRKDLQN 100
#> 22:    0.167 7.5 0.0218844453 198.5254      5 LVRKDLQN 100
#> 23:    0.167 7.5 0.0040049980 198.7266      5 LVRKDLQN 100
#> 24:    0.167 7.5 0.0006199309 198.9279      5 LVRKDLQN 100
#> 25:    5.000 7.5 0.3690002996 329.1978      3 LVRKDLQN 100
#> 26:    5.000 7.5 0.3680489595 329.5332      3 LVRKDLQN 100
#> 27:    5.000 7.5 0.1836859861 329.8687      3 LVRKDLQN 100
#> 28:    5.000 7.5 0.0606794961 330.2041      3 LVRKDLQN 100
#> 29:    5.000 7.5 0.0150560010 330.5395      3 LVRKDLQN 100
#> 30:    5.000 7.5 0.0030169400 330.8749      3 LVRKDLQN 100
#> 31:    5.000 7.5 0.0004630632 331.2104      3 LVRKDLQN 100
#> 32:    5.000 7.5 0.3690002996 197.9216      5 LVRKDLQN 100
#> 33:    5.000 7.5 0.3680489595 198.1229      5 LVRKDLQN 100
#> 34:    5.000 7.5 0.1836859861 198.3241      5 LVRKDLQN 100
#> 35:    5.000 7.5 0.0606794961 198.5254      5 LVRKDLQN 100
#> 36:    5.000 7.5 0.0150560010 198.7266      5 LVRKDLQN 100
#> 37:    5.000 7.5 0.0030169400 198.9279      5 LVRKDLQN 100
#> 38:    5.000 7.5 0.0004630632 199.1291      5 LVRKDLQN 100
#>     Exposure  PH    Intensity       Mz Charge Sequence  PF