HDX-Analyzer model

This function performs the test based on the simplest linear models for deuteration curves containing time, state of the protein and the interaction term. Its input and output are compatible with the function calculate_hdx_power.

test_hdx_analyzer(data, significance_level = 0.05)

Arguments

data

data.table with deuteration curves

significance_level

significance level for tests

Value

This function returns a data table compatible with the function calculate_hdx_power.

References

Liu, Sanmin et al. (2011). “HDX-analyzer: a novel package for statistical analysis of protein structure dynamics”. In:BMC bioinformatics12.1, pp. 1–10.

See also

Examples

theo_spectra_pf_100 <- simulate_theoretical_spectra(sequence = "LVRKDLQN",
                                                    charge = c(3, 5),
                                                    protection_factor = 100,
                                                    times = c(0.167, 5),
                                                    pH = 7.5,
                                                    temperature = 15,
                                                    n_molecules = 500,
                                                    time_step_const = 1,
                                                    use_markov = TRUE)
theo_spectra_pf_200 <- simulate_theoretical_spectra(sequence = "LVRKDLQN",
                                                    charge = c(3, 5),
                                                    protection_factor = 200,
                                                    times = c(0.167, 5),
                                                    pH = 7.5,
                                                    temperature = 15,
                                                    n_molecules = 500,
                                                    time_step_const = 1,
                                                    use_markov = TRUE)

theo_spectra_two_states <- rbind(theo_spectra_pf_100, theo_spectra_pf_200)

deut_curves_p_states <- get_noisy_deuteration_curves(theo_spectra_two_states,
                                                     n_replicates = 4,
                                                     n_experiments = 1,
                                                     reference = 100)[[1]][[1]]
test_hdx_analyzer(deut_curves_p_states)
#>                                Test State_1 State_2 Test_statistic    P_value
#> 1:  Deuteros lm identity continuous     200     100       1.280723 0.28798967
#> 2: Deuteros lm identity categorical     200     100       3.171940 0.03391475
#> 3:       Deuteros lm log continuous     200     100       1.390142 0.25976237
#>    Significant_difference        Time Transformation       AIC     logLik
#> 1:                  FALSE  continuous       identity  15.72927 -2.8646370
#> 2:                   TRUE categorical       identity -44.12452 29.0622609
#> 3:                  FALSE  continuous            log  10.99769 -0.4988453