Hydrogen-deuterium or back-exchange exchange rates

Calculate exchange rates that are required to obtain exchange probabilities.

get_exchange_rates(
  sequence,
  exchange = "HD",
  pH = 9,
  temperature = 15,
  mol_type = "poly",
  if_corr = FALSE
)

Arguments

sequence

peptide amino acid sequence as a character vector of amino acids

exchange

type of exchange - "HD" for hydrogen to deuterium, "DH" for deuterium to hydrogen (back-exchange). Default "HD".

pH

reaction pH

temperature

temperature of the reaction (Celsius). Default to 15.

mol_type

character, "poly" or "oligo". If "oligo" then the calculated constants are multiplied by c according to the considered condition (base, water or acid).

if_corr

logical. PH correction indicator. Default value FALSE. The value of pH is equal to pD. If there is correction, the pD = pH + 0.4. (Conelly et al 1993)

Value

The obtained exchange rates are stored in vector kcHD or kcDH according to the exchange direction. They are used to calculate the exchange probabilities thus both kcHD and kcDH are necessary as we take the possibility of back-exchange into account.

Details

The correction of pH is taken into account for calculation of pD: $$pD = pH + 0.4 * if_corr$$ Next, the provided temperature is converted into K and the internal functions get_F_const, get_poly_const and get_pkc are evaluated.

Using the obtained matrix of constants and provided sequence F_a and F_b are calculated for each amino acid in the sequence, concerning the previous and next amino acid. For the amino acids in the middle of the sequence, the following formula is used:

$$F_x = 10^{ previous_x + current_x}$$

where x is either a or b, and previous_x is the acid/base factor for a previous amino acid in the sequence, and current_x for the amino acid it is calculated for. If the amino acid is next to the C- or N-term, the term-effect is taken into account.

Finally, the exchange rate k_c for the amino acid is the sum of catalysis constants for acid, base and water (Conelly et al, 1993). Namely:

$$k_c = k_{acid} + k_{base} + k_{water}$$ where $$k_{acid} = F_a * K_a + D * F_ta,$$ $$k_{base} = F_b* K_b + OD * F_tb,$$ $$k_{water} = F_b * K_w * F_tw$$

where D and OD indicates deuterium and deuterium oxide concentration, F_a and F_b are values calculated specifically for given amino acid, as described before, K_a and K_b are values computed by get_poly_const function, based on the mole type, F_ta, F_tb and F_tw are values computed by get_F_const function.