Semiparametric test for differences in deuteration levels

This function performs the semiparametric test for differences in deuteration levels. Its input and output are compatible with the function calculate_hdx_power.

test_semiparametric(data, significance_level = 0.05)

Arguments

data
data.table with deuteration curves

significance_level
significance level for tests

Value

This function returns a data table compatible with the function calculate_hdx_power.

Details

This function uses truncated_lines. The knots considered in the testing procedure are chosen using ridge regression.

See also

Examples

theo_spectra_pf_100 <- simulate_theoretical_spectra(sequence = "LVRKDLQN",
                                                    charge = c(3, 5),
                                                    protection_factor = 100,
                                                    times = c(0.167, 5, 10, 30),
                                                    pH = 7.5,
                                                    temperature = 15,
                                                    n_molecules = 500,
                                                    time_step_const = 1,
                                                    use_markov = TRUE)
theo_spectra_pf_200 <- simulate_theoretical_spectra(sequence = "LVRKDLQN",
                                                    charge = c(3, 5),
                                                    protection_factor = 200,
                                                    times = c(0.167, 5, 10, 30),
                                                    pH = 7.5,
                                                    temperature = 15,
                                                    n_molecules = 500,
                                                    time_step_const = 1,
                                                    use_markov = TRUE)

theo_spectra_two_states <- rbind(theo_spectra_pf_100, theo_spectra_pf_200)

deuteration_curves_paired_states <- get_noisy_deuteration_curves(theo_spectra_two_states,
                                                                 n_replicates = 4,
                                                                 n_experiments = 1,
                                                                 reference = 100)[[1]][[1]]
test_semiparametric(deuteration_curves_paired_states)

#> fixed-effect model matrix is rank deficient so dropping 1 column / coefficient

#> boundary (singular) fit: see ?isSingular

#> fixed-effect model matrix is rank deficient so dropping 1 column / coefficient

#> boundary (singular) fit: see ?isSingular

#>                                        Test State_1 State_2 Test_statistic
#> 1: RIDGE_knots_random_intercept_id_exposure     200     100       5.820557
#>       P_value Significant_difference Time Transformation       AIC   logLik
#> 1: 0.05446056                  FALSE   NA       identity -79.53788 49.76894